+++ src/task/task.service.ts
@@ -1,14 +1,54 @@
+import { ProjectService } from './../project/project.service';
 import { DeleteTaskDto } from './dto/delete-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateTaskDto } from './dto/create-task.dto';
+import { ColumnService } from 'src/column/column.service';
 
 @Injectable()
 export class TaskService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly columnService: ColumnService,
+    private readonly projectService: ProjectService,
+  ) {}
 
-  async updateTask(task: UpdateTaskDto) {
+  async createTask(userId: number, task: CreateTaskDto) {
+    const column = await this.columnService.findColumnByIdWithProject(
+      task.columnId,
+    );
+
+    if (!column) throw new BadRequestException('There is no such column');
+
+    if (!this.projectService.isProjectMember(userId, column.project))
+      throw new BadRequestException('The user is not a member');
+
+    const order = column.tasks.reduce(
+      (acc, col) => (acc > col.order ? acc : col.order + 1),
+      1,
+    );
+
+    const createdTask = await this.prisma.task.create({
+      data: {
+        ...task,
+        order,
+      },
+    });
+
+    return createdTask;
+  }
+
+  async updateTask(userId: number, task: UpdateTaskDto) {
+    const column = await this.columnService.findColumnByIdWithProject(
+      task.columnId,
+    );
+
+    if (!column) throw new BadRequestException('There is no such column');
+
+    if (!this.projectService.isProjectMember(userId, column.project))
+      throw new BadRequestException('The user is not a member');
+
     const updatedTask = await this.prisma.task.update({
       data: {
         title: task.title,
@@ -23,7 +63,16 @@ export class TaskService {
     return updatedTask;
   }
 
-  async deleteTask(task: DeleteTaskDto) {
+  async deleteTask(userId: number, task: DeleteTaskDto) {
+    const column = await this.columnService.findColumnByIdWithProject(
+      task.columnId,
+    );
+
+    if (!column) throw new BadRequestException('There is no such column');
+
+    if (!this.projectService.isProjectMember(userId, column.project))
+      throw new BadRequestException('The user is not a member');
+
     const deletedTask = await this.prisma.task.delete({
       where: {
         id: task.id,
@@ -53,29 +102,5 @@ export class TaskService {
     );
 
     return await this.prisma.$transaction(updatedTasks);
-  }
-
-  async addTaskToProject(task: CreateTaskDto) {
-    const column = await this.prisma.column.findFirst({
-      where: {
-        id: task.columnId,
-      },
-      include: {
-        tasks: true,
-      },
-    });
-
-    const order = column.tasks.reduce(
-      (acc, col) => (acc > col.order ? acc : col.order + 1),
-      1,
-    );
-
-    return await this.prisma.task.create({
-      data: {
-        title: task.title,
-        columnId: column.id,
-        order: order,
-      },
-    });
   }
 }
